# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sentinel file intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /action-queue:
    get:
      operationId: listActionQueue
      tags: [review]
      summary: List proposed (never-executed) file operations awaiting a future explicit trigger
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, dismissed]
      responses:
        "200":
          description: Action queue
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ActionQueueEntry"
  /action-queue/{id}/dismiss:
    patch:
      operationId: dismissActionQueueItem
      tags: [review]
      summary: Dismiss a queued action without ever performing it
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dismissed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionQueueEntry"
        "404":
          description: Not found
...
  /findings/{id}/extract:
    post:
      operationId: extractFindingText
      tags: [extraction]
      summary: Run text extraction (and OCR if applicable/enabled) for one finding, on demand
      description: >
        Extraction is always per-file and on demand — there is no bulk/automatic mode.
        Fails with 409 if extraction is disabled in Settings for the required mode
        (OCR / cloud consent).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Extraction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractedTextEntry"
        "404":
          description: Finding not found
        "409":
          description: Extraction/OCR disabled in settings
  /findings/{id}/extraction:
    get:
      operationId: getFindingExtraction
      tags: [extraction]
      summary: Latest extracted text + detected entities + sensitive-content flags for a finding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Extraction detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractionDetailResponse"
        "404":
          description: No extraction on file for this finding
  /findings/{id}/summarize:
    post:
      operationId: summarizeFinding
      tags: [extraction]
      summary: Generate an AI summary for one explicitly selected document (requires cloud consent if a cloud provider is used)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SummaryResponse"
        "404":
          description: No extracted text to summarize
        "409":
          description: Cloud consent required
components:
  schemas:
...
    ActionQueueEntry:
      type: object
      required: [id, findingId, actionType, description, status, createdAt]
      properties:
        id:
          type: integer
        findingId:
          type: integer
        actionType:
          type: string
          enum: [move, delete, archive, rename]
        proposedDestination:
          type: ["string", "null"]
        description:
          type: string
        status:
          type: string
          enum: [pending, dismissed]
        createdAt:
          type: string
...
    ExtractedTextEntry:
      type: object
      required: [id, findingId, extractor, text, truncated, sensitiveCategories, createdAt]
      properties:
        id:
          type: integer
        findingId:
          type: integer
        extractor:
          type: string
          enum: [pdf, txt, csv, json, markdown, source_code, ocr]
        text:
          type: string
        truncated:
          type: boolean
        sensitiveCategories:
          type: array
          items:
            type: string
            enum: [legal, banking, medical, identity, api_key, password, private_key]
        ocrProvider:
          type: ["string", "null"]
        createdAt:
          type: string
    EntityEntry:
      type: object
      required: [id, findingId, type, value, createdAt]
      properties:
        id:
          type: integer
        findingId:
          type: integer
        type:
          type: string
          enum: [person, organization, date, invoice_number, case_reference, amount]
        value:
          type: string
        createdAt:
          type: string
    ExtractionDetailResponse:
      type: object
      required: [extraction, entities]
      properties:
        extraction:
          $ref: "#/components/schemas/ExtractedTextEntry"
        entities:
          type: array
          items:
            $ref: "#/components/schemas/EntityEntry"
    SummaryResponse:
      type: object
      required: [findingId, summary, provider, requiresCloudConsent]
      properties:
        findingId:
          type: integer
        summary:
          type: string
        provider:
          type: string
        requiresCloudConsent:
          type: boolean