# lib/api-zod/src/generated/types/entityEntry.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sentinel file intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { EntityEntryType } from './entityEntryType';

export interface EntityEntry {
  id: number;
  findingId: number;
  type: EntityEntryType;
  value: string;
  createdAt: string;
}