# lib/api-zod/src/generated/types/extractedTextEntry.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sentinel file intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { ExtractedTextEntryExtractor } from './extractedTextEntryExtractor';
import type { ExtractedTextEntrySensitiveCategoriesItem } from './extractedTextEntrySensitiveCategoriesItem';

export interface ExtractedTextEntry {
  id: number;
  findingId: number;
  extractor: ExtractedTextEntryExtractor;
  text: string;
  truncated: boolean;
  sensitiveCategories: ExtractedTextEntrySensitiveCategoriesItem[];
  /** @nullable */
  ocrProvider?: string | null;
  createdAt: string;
}